-added a resize scale to each frame func resize to resize stored image pil objects
-changed
structure of  class Path's _dict to generate beforehand the full path of each image, and total count.

on a row selection the artioned _dict is not sent in a function call, but injected into the Gallery instance with uniform key naming scheme.

func getcoords to give a specific generator based on an explicit string 'demand'.